[MouseWithoutBorders] - moving Common.Helper.cs -> Core\Helper.cs - #35155
